FoundryVTT VM improvements and pihole VM setup
Technitium and its presence in the k3s cluster has proven to be a major PITA for circular dependencies, and then it fell over because its disk was full. It was not worth restoring.

These changes switch over to pihole as a VM (along with unbound) and also a few other changes that have accumulated. technitium is not yet deprovisioned.